profile refactor
