+++ src/client/pages/profile/queries.js
@@ -1,0 +1,16 @@
+import gql from "graphql-tag";
+
+export const GET_PROFILE = gql`
+  query profile($email: String!) {
+    profile(email: $email) {
+      email
+      name
+      location
+      workType
+      skills
+      tagline
+      statement
+      experience
+    }
+  }
+`;
